+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Download current and previous releases from [https://github.com/ludus-leonis/Fre
 * added tabular numbers for library editor tree/preview
 * added JSDoc linting and moved linting from gulp to npm
 * added GitHub Actions for automated build
+* added error message when snapshot list is empty
 * fixed texture bug on masked non-square assets
 * fixed click on icon in popup does not trigger function
 * changed new-note to show the edit window first instead of creating an empty note